Accept Integer for accelerator backend priority on 32-bit.
`priority=100` is Int32 on i686 while register_backend required Int64,
breaking Reactant.__init__ (and dependents like NeuralOperators).

+++ src/accelerators/CPU.jl
@@ -49,7 +49,7 @@ end
 function __init__()
     register_backend(
         "cpu";
-        priority=100,
+        priority=Int64(100),
         pjrt_initialize_function=make_pjrt_client,
         ifrt_initialize_function=make_ifrt_client,
     )
