+++ src/accelerators/TT.jl
@@ -73,7 +73,7 @@ function __init__()
     if Sys.islinux() && has_tt() && !Reactant.precompiling()
         register_backend(
             "tt";
-            priority=1000,
+            priority=Int64(1000),
             pjrt_initialize_function=make_pjrt_client,
             ifrt_initialize_function=make_ifrt_client,
             preinitialize_setup_function=() -> begin
