+++ src/accelerators/Trainium.jl
@@ -401,7 +401,7 @@ function __init__()
     if Sys.islinux() && has_trainium() && !Reactant.precompiling()
         register_backend(
             "trainium";
-            priority=1000,
+            priority=Int64(1000),
             pjrt_initialize_function=make_pjrt_client,
             ifrt_initialize_function=make_ifrt_client,
             preinitialize_setup_function=() -> begin
